+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/CredentialRequestProofSupported.kt
@@ -12,5 +12,12 @@ data class CredentialRequestProofSupported(
      */
     @SerialName("proof_signing_alg_values_supported")
     val supportedSigningAlgorithms: Collection<String>,
-    // todo add key_attestations_required
+
+    /**
+     * OID4VCI: OPTIONAL. Object that describes the requirement for key attestations as described in Appendix D,
+     * which the Credential Issuer expects the Wallet to send within the [CredentialRequestProof].
+     * If the Credential Issuer does not require a key attestation, this parameter MUST NOT be present in the metadata.
+     */
+    @SerialName("key_attestations_required")
+    val keyAttestationRequired: KeyAttestationRequired? = null,
 )